# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Jeff Buitrago artist website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
}

export interface Show {
  id: number;
  place: string;
  day: string;
  /** @nullable */
  time: string | null;
  available: boolean;
  createdAt: string;
}

export interface CreateShowBody {
  place: string;
  day: string;
  /** @nullable */
  time?: string | null;
  available?: boolean;
}

export interface UpdateShowBody {
  place?: string;
  day?: string;
  /** @nullable */
  time?: string | null;
  available?: boolean;
}

export interface Photo {
  id: number;
  url: string;
  title: string;
  /** @nullable */
  titleEn: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreatePhotoBody {
  url: string;
  title: string;
  /** @nullable */
  titleEn?: string | null;
  sortOrder?: number;
}

export interface UpdatePhotoBody {
  url?: string;
  title?: string;
  /** @nullable */
  titleEn?: string | null;
  sortOrder?: number;
}

export interface Video {
  id: number;
  url: string;
  title: string;
  /** @nullable */
  titleEn: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreateVideoBody {
  url: string;
  title: string;
  /** @nullable */
  titleEn?: string | null;
  sortOrder?: number;
}

export interface UpdateVideoBody {
  url?: string;
  title?: string;
  /** @nullable */
  titleEn?: string | null;
  sortOrder?: number;
}