# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Jeff Buitrago artist website
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * @summary Admin logout
 */
export const AdminLogoutResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * @summary Get current auth status
 */
export const GetAuthMeResponse = zod.object({
  authenticated: zod.boolean(),
});

/**
 * @summary List all shows
 */
export const ListShowsResponseItem = zod.object({
  id: zod.number(),
  place: zod.string(),
  day: zod.string(),
  time: zod.string().nullable(),
  available: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListShowsResponse = zod.array(ListShowsResponseItem);

/**
 * @summary Create a show (admin only)
 */
export const CreateShowBody = zod.object({
  place: zod.string(),
  day: zod.string(),
  time: zod.string().nullish(),
  available: zod.boolean().optional(),
});

/**
 * @summary Update a show (admin only)
 */
export const UpdateShowParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateShowBody = zod.object({
  place: zod.string().optional(),
  day: zod.string().optional(),
  time: zod.string().nullish(),
  available: zod.boolean().optional(),
});

export const UpdateShowResponse = zod.object({
  id: zod.number(),
  place: zod.string(),
  day: zod.string(),
  time: zod.string().nullable(),
  available: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a show (admin only)
 */
export const DeleteShowParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all photos
 */
export const ListPhotosResponseItem = zod.object({
  id: zod.number(),
  url: zod.string(),
  title: zod.string(),
  titleEn: zod.string().nullable(),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListPhotosResponse = zod.array(ListPhotosResponseItem);

/**
 * @summary Add a photo (admin only)
 */
export const CreatePhotoBody = zod.object({
  url: zod.string(),
  title: zod.string(),
  titleEn: zod.string().nullish(),
  sortOrder: zod.number().optional(),
});

/**
 * @summary Update a photo (admin only)
 */
export const UpdatePhotoParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePhotoBody = zod.object({
  url: zod.string().optional(),
  title: zod.string().optional(),
  titleEn: zod.string().nullish(),
  sortOrder: zod.number().optional(),
});

export const UpdatePhotoResponse = zod.object({
  id: zod.number(),
  url: zod.string(),
  title: zod.string(),
  titleEn: zod.string().nullable(),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a photo (admin only)
 */
export const DeletePhotoParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all videos
 */
export const ListVideosResponseItem = zod.object({
  id: zod.number(),
  url: zod.string(),
  title: zod.string(),
  titleEn: zod.string().nullable(),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListVideosResponse = zod.array(ListVideosResponseItem);

/**
 * @summary Add a video (admin only)
 */
export const CreateVideoBody = zod.object({
  url: zod.string(),
  title: zod.string(),
  titleEn: zod.string().nullish(),
  sortOrder: zod.number().optional(),
});

/**
 * @summary Update a video (admin only)
 */
export const UpdateVideoParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateVideoBody = zod.object({
  url: zod.string().optional(),
  title: zod.string().optional(),
  titleEn: zod.string().nullish(),
  sortOrder: zod.number().optional(),
});

export const UpdateVideoResponse = zod.object({
  id: zod.number(),
  url: zod.string(),
  title: zod.string(),
  titleEn: zod.string().nullable(),
  sortOrder: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a video (admin only)
 */
export const DeleteVideoParams = zod.object({
  id: zod.coerce.number(),
});