# lib/api-zod/src/generated/types/createVideoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Jeff Buitrago artist website
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVideoBody {
  url: string;
  title: string;
  /** @nullable */
  titleEn?: string | null;
  sortOrder?: number;
}