# lib/api-zod/src/generated/types/show.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Jeff Buitrago artist website
 * OpenAPI spec version: 0.1.0
 */

export interface Show {
  id: number;
  place: string;
  day: string;
  /** @nullable */
  time: string | null;
  available: boolean;
  createdAt: Date;
}